+++ indexer/integration_test_util/db_suite_read_record.go
@@ -112,7 +112,8 @@ SELECT
 	partition_id, -- 4
 	epoch, -- 5
 	message_types, -- 6
-	tx_type -- 7
+	tx_type, -- 7
+	"action" -- 8
 FROM "transaction" WHERE hash = $1 AND height = $2 AND chain_id = $3
 `
 
@@ -140,6 +141,7 @@ FROM "transaction" WHERE hash = $1 AND height = $2 AND chain_id = $3
 		&res.Epoch,                  // 5
 		pq.Array(&res.MessageTypes), // 6
 		&res.TxType,                 // 7
+		&res.Action,                 // 8
 	)
 	suite.Require().NoError(err, "failed to scan transaction")
 	suite.Require().Falsef(rows.Next(), "more than one record found for transaction %s at %d", hash, height)
@@ -158,7 +160,8 @@ SELECT
 	hash, -- 3
 	ref_count, -- 4
 	epoch, -- 5
-	message_types -- 6
+	message_types, -- 6
+	"action" -- 7
 FROM recent_account_transaction WHERE hash = $1 AND height = $2 AND chain_id = $3
 `
 
@@ -185,6 +188,7 @@ FROM recent_account_transaction WHERE hash = $1 AND height = $2 AND chain_id = $
 		&res.RefCount,               // 4
 		&res.Epoch,                  // 5
 		pq.Array(&res.MessageTypes), // 6
+		&res.Action,                 // 7
 	)
 	suite.Require().NoError(err, "failed to scan recent account transaction")
 	suite.Require().Falsef(rows.Next(), "more than one record found for recent account transaction %s at %d", hash, height)